feat(JAQPOT-4): add user id on model

+++ src/main/kotlin/org/jaqpot/api/repository/ModelRepository.kt
@@ -3,4 +3,4 @@ package org.jaqpot.api.repository
 import org.jaqpot.api.entity.Model
 import org.springframework.data.repository.CrudRepository
 
-interface ModelRepository : CrudRepository<Model, Int>
+interface ModelRepository : CrudRepository<Model, Long>
